Make AWS KMS an optional feature (via install extras)
For AWS KMS support we pull in the `boto3` package, which has a lot of
dependencies that users don't need unless they use AWS. In addition, the
`boto3` and `botocore` packages update their version numbers almost on a
daily basis, which is a huge nightmare when you track dependencies e.g.
via automatic generation of `requirements.txt`.